Non-admin users should no longer be able to see the link to "Manage Users", nor should they be able to see links to delete or promote other users.
Added unit tests for users and for all validations and methods.

Added static_pages controller for "home", "about", and "help" pages. These should be accessible through "~/home", etc. as defined in routes.rb. The root url is now "~/home", which provides a splash page for what the app is about and how to use it. Also added our team information to the About page.